+++ DigitalOcean.API/Clients/IImageActionsClient.cs
@@ -6,16 +6,16 @@ namespace DigitalOcean.API.Clients {
         /// <summary>
         /// Transfer an Image to another region
         /// </summary>
-        Task<Action> Transfer(int imageId, string regionSlug);
+        Task<Action> Transfer(long imageId, string regionSlug);
 
         /// <summary>
         /// To convert an image, for example, a backup to a snapshot.
         /// </summary>
-        Task<Action> Convert(int imageId);
+        Task<Action> Convert(long imageId);
 
         /// <summary>
         /// Retrieve an existing Image Action
         /// </summary>
-        Task<Action> GetAction(int imageId, int actionId);
+        Task<Action> GetAction(long imageId, long actionId);
     }
 }
