+++ DigitalOcean.API/Models/Requests/Firewall.cs
@@ -25,7 +25,7 @@ namespace DigitalOcean.API.Models.Requests {
         /// An array containing the IDs of the Droplets to be assigned to the Firewall.
         /// </summary>
         [JsonProperty("droplet_ids")]
-        public List<int> DropletIds { get; set; }
+        public List<long> DropletIds { get; set; }
 
         /// <summary>
         /// An array containing the names of the Tags to be assigned to the Firewall.
